Added boundary between UI and Operation.
